moved HasSelection to com.vaadin.flow.component
